Refactor spot type

See merge request s10-final/S10P31E106!87

+++ backend/connectus/spot-service/src/main/java/social/connectus/location/common/type/Position.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PUBLIC)
 @AllArgsConstructor(access = AccessLevel.PUBLIC)
 public class Position implements Serializable {
-	private double longitude;
-	private double latitude;
+	private float longitude;
+	private float latitude;
 }
